♻️ refactor(webapp): extract permissions settings panels
